fix stats decremented with unsigned values
Reported by Coverity CID #200104, #200030, #200096, #200091, #200090,
 #200074, #200073, #200049, #200039, #200035, #200032, #200009,
 #200003, #199998, #199996, #199991, #199988, #199986, #199985,
 #199982, #199977, #199970, #199964, #199963, #199960, #199952,
 #199948, #199945, #199941, #199938, #199934, #199923, #199920
 #199913, #199900, #199896

+++ mem/hp_malloc_stats.c
@@ -225,9 +225,9 @@ void hp_init_rpm_statistics(struct hp_block *hpb)
 	rpm_rused->flags |= STAT_NO_RESET;
 	rpm_frags->flags |= STAT_NO_RESET;
 #endif
-	update_stat(rpm_used, hpb->used);
-	update_stat(rpm_rused, hpb->real_used);
-	update_stat(rpm_frags, hpb->total_fragments);
+	update_stat(rpm_used, (int)hpb->used);
+	update_stat(rpm_rused, (int)hpb->real_used);
+	update_stat(rpm_frags, (int)hpb->total_fragments);
 
 	LM_DBG("initializing atomic rpm statistics: "
 	       "[ us: %ld | rus: %ld | frags: %ld ]\n", hpb->used, hpb->real_used, hpb->total_fragments);
